Added test cases for orders API endpoints - GET view orders, POST create order, POST add/update items, fixtures, and warning handling
